move load_module to util
